+++ src/main/java/org/torproject/metrics/onionoo/writer/BandwidthDocumentWriter.java
@@ -64,25 +64,19 @@ public class BandwidthDocumentWriter implements DocumentWriter {
   }
 
   private String[] graphNames = new String[] {
-      "3_days",
-      "1_week",
       "1_month",
       "6_months",
       "1_year",
       "5_years" };
 
   private Period[] graphIntervals = new Period[] {
-      Period.ofDays(3),
-      Period.ofWeeks(1),
       Period.ofMonths(1),
       Period.ofMonths(6),
       Period.ofYears(1),
       Period.ofYears(5) };
 
   private long[] dataPointIntervals = new long[] {
-      DateTimeHelper.FIFTEEN_MINUTES,
-      DateTimeHelper.ONE_HOUR,
-      DateTimeHelper.FOUR_HOURS,
+      DateTimeHelper.ONE_DAY,
       DateTimeHelper.ONE_DAY,
       DateTimeHelper.TWO_DAYS,
       DateTimeHelper.TEN_DAYS };