+++ src/main/java/org/torproject/metrics/onionoo/writer/ClientsDocumentWriter.java
@@ -78,16 +78,19 @@ public class ClientsDocumentWriter implements DocumentWriter {
   }
 
   private String[] graphNames = new String[] {
+      "1_month",
       "6_months",
       "1_year",
       "5_years" };
 
   private Period[] graphIntervals = new Period[] {
+      Period.ofMonths(1),
       Period.ofMonths(6),
       Period.ofYears(1),
       Period.ofYears(5) };
 
   private long[] dataPointIntervals = new long[] {
+      DateTimeHelper.ONE_DAY,
       DateTimeHelper.ONE_DAY,
       DateTimeHelper.TWO_DAYS,
       DateTimeHelper.TEN_DAYS };